QA-4760 step  added after - before
